Merge release/v3.2.0 to main

* commit 'd880e10e538160f643593aeb017d908acb60f31b': (93 commits)
  Update v3.2.0 Debian changelog
  Clean up parse diagnostic documentation
  fixed ->'s added by sed, and depricated parse_debug_ex reference in comment
  fixed tests diagnostic leaks
  removed system allocator hard code
  add unsaved change
  added h_parse_diagnostic_trace_fprint to view the full execution trace
  ran clang-format
  added hammer assertions
  resolve leftover merge conflict
  removed needless type conversion
  apply fixes and formatting
  Address UBSan review findings
  fix sign conversion errors
  Fix UBSan CI build
  Run CI with undefined behavior sanitizer
  Fix undefined behavior sanitizer issues
  Bump version to 3.2.0
  combined h_parse_debug_ex and h_parse_debug into one function
  Apply clang-format to error reporting changes
  ...

+++ src/cfgrammar.h
@@ -104,11 +104,11 @@ const HStringMap *h_follow(size_t k, HCFGrammar *g, const HCFChoice *x);
 HStringMap *h_predict(size_t k, HCFGrammar *g, const HCFChoice *A, const HCFSequence *rhs);
 
 /* Pretty-printers for grammars and associated data. */
-void h_pprint_grammar(FILE *file, const HCFGrammar *g, int indent);
+void h_pprint_grammar(FILE *file, const HCFGrammar *g, size_t indent);
 void h_pprint_sequence(FILE *f, const HCFGrammar *g, const HCFSequence *seq);
 void h_pprint_symbol(FILE *f, const HCFGrammar *g, const HCFChoice *x);
-void h_pprint_symbolset(FILE *file, const HCFGrammar *g, const HHashSet *set, int indent);
-void h_pprint_stringset(FILE *file, const HStringMap *set, int indent);
+void h_pprint_symbolset(FILE *file, const HCFGrammar *g, const HHashSet *set, size_t indent);
+void h_pprint_stringset(FILE *file, const HStringMap *set, size_t indent);
 void h_pprint_stringmap(FILE *file, char sep, void (*valprint)(FILE *f, void *env, void *val),
                         void *env, const HStringMap *map);
 void h_pprint_char(FILE *file, uint8_t c);
